Refactor Toppler game setup for ESM and improve HTML generation
